refactor: (#95) Main ROOM 탭 상수/유틸 분리
